before rebase with master

+++ srcs/Client/Client.accessor.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Client.accessor.cpp                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rgeny <rgeny@student.42.fr>                +#+  +:+       +#+        */
+/*   By: ayzapata <ayzapata@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/12 14:56:10 by rgeny             #+#    #+#             */
-/*   Updated: 2022/04/20 05:09:01 by rgeny            ###   ########.fr       */
+/*   Updated: 2022/04/24 15:01:42 by ayzapata         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ bool	Client::set_nickname	(std::string nickname)
 	{
 		return (false);
 	}
-	for (int i = 1; i < 8 && i < nickname.size(); i++)
+	for (size_t i = 1; i < 8 && i < nickname.size(); i++)
 	{
 		c = nickname[i];
 		if (!isalnum(c)
